add support for map type: ROADMAP, SATELLITE
I took the current version, v1.3 from the WordPress repo and modified the plugin to support a map type. Values are 'roadmap' or 'satellite'. It defaults to ROADMAP just like the v1.3.

The GitHub repo I forked it from wasn't up to date either. This brings it up to v1.3 in the repo and adds the new feature.